+++ src/talos/persistence/models/__init__.py
@@ -10,6 +10,7 @@ from talos.persistence.models.asin_master import AsinMaster
 from talos.persistence.models.cart_item import CartItem
 from talos.persistence.models.config_override import ConfigOverride
 from talos.persistence.models.listino_item import ListinoItem
+from talos.persistence.models.locked_in_item import LockedInItem
 from talos.persistence.models.panchina_item import PanchinaItem
 from talos.persistence.models.storico_ordine import StoricoOrdine
 from talos.persistence.models.vgp_result import VgpResult
@@ -20,6 +21,7 @@ __all__ = [
     "CartItem",
     "ConfigOverride",
     "ListinoItem",
+    "LockedInItem",
     "PanchinaItem",
     "StoricoOrdine",
     "VgpResult",